Back/Django/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
joblib==1.2.0
jsonschema==4.17.3
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
promise==2.3
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
text-unidecode==1.3
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.15
